fix: make CWs dependent upon 'spoiler_text' and not 'sentitive' value
zirk.us for example seems to mark all new toots with "sensitive: true",
which causes a CW to show for every toot.

Since it is not possible to post a toot with an empty CW field, I can
safely assume that any text in the CW field means there is a CW and just
ignore the 'sensitive' field.

fixes: #59