feat: switch from google-closure-compler to rollup & terser
I did not really want to use gcc in the first place as it felt to
heavy-wight and resource intensive. But I was forced to because I could
not get any other solution to compile js correctly.

However because some resource gcc needs is not available on my
webserver I was still unable to use it as intended remotely.

By combining rollup and terser I was now able to combine and minify all
the separate js files correctly.

This also removes the checked-in compiled site from the repo.