build: migrate to node application for serving and building
Since my hosting package does not allow me to run npm commands as part
of a post-checkout step I was forced to manually run the build command
for each new commit.

This commit migrates to using an express application for serving the
site, and also integrates running the 'build' command for every update.

fixes: #69